Adding GameUtil. Updating Duo and Grid match pages.
